pyproject.toml
[project]
name = "ml-way"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "mysql-connector-python>=9.2.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
]
